// apps/web/src/api/model/tocNodeResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Scholia API
 * OpenAPI spec version: 0.1.0
 */

export interface TocNodeResponse {
    children: TocNodeResponse[];
    depth: number;
    has_content: boolean;
    id: string;
    label: string;
    label_html?: string;
    slug: string;
    sort_order: number;
    source_id?: string;
    source_node_id?: string;
    source_ref: string;
}